Fixed remove argument

+++ src/arguments.c
@@ -1,9 +1,10 @@
-#include "../include/utils.h"
+#include "utils.h"
 
 struct stat buffer;
 
 void set_font(const char *font_name) {
     char command[200];
+
     snprintf(command, sizeof(command), "cp " FONTS_DIRECTORY "/%s.ttf " FONT_SET_DEFAULT, font_name);
     system(command);
     system("termux-reload-settings");
@@ -11,6 +12,14 @@ void set_font(const char *font_name) {
 
 void font_remove(const char *font_removing) {
   char command[200];
+
+  if (!font_exists(font_removing)) {
+    printf("no as agregador ninguna fuente aqui");
+    return;
+  }
+
   snprintf(command, sizeof(command), "rm " FONTS_DIRECTORY "/%s.ttf", font_removing);
   system(command);
+  system("termux-reload-settings");
+  
 }
